+++ src/Shell.tsx
@@ -1,10 +1,10 @@
-import { User } from 'botframework-directlinejs';
 import * as React from 'react';
-import { connect, Dispatch } from 'react-redux';
+import { connect } from 'react-redux';
+
 import { classList } from './Chat';
 import { Speech } from './SpeechModule';
-import { ChatState, FormatState } from './Store';
 import { ChatActions, ListeningState, sendFiles, sendMessage } from './Store';
+import { ChatState } from './Store';
 import { Strings } from './Strings';
 
 interface Props {
@@ -39,7 +39,10 @@ class ShellContainer extends React.Component<Props> implements ShellFunctions {
         if (evt.key === 'Enter' || evt.key === ' ') {
             evt.preventDefault();
             this.sendMessage();
-            this.textInput.focus();
+
+            if (this.textInput) {
+                this.textInput.focus();
+            }
         }
     }
 
@@ -67,7 +70,9 @@ class ShellContainer extends React.Component<Props> implements ShellFunctions {
             this.fileInput.value = null;
         }
 
-        this.textInput.focus();
+        if (this.textInput) {
+            this.textInput.focus();
+        }
     }
 
     private onTextInputFocus() {
@@ -85,7 +90,9 @@ class ShellContainer extends React.Component<Props> implements ShellFunctions {
     }
 
     public focus(appendKey?: string) {
-        this.textInput.focus();
+        if (this.textInput) {
+            this.textInput.focus();
+        }
 
         if (appendKey) {
             this.props.onChangeText(this.props.inputText + appendKey);